Modularize remoteDB tests
