+++ app/src/main/AndroidManifest.xml
@@ -41,7 +41,7 @@
             android:label="@string/title_activity_greeting"
             android:theme="@style/Theme.Polypoly" />
         <activity
-            android:name=".WelcomeActivity"
+            android:name=".menu.WelcomeActivity"
             android:exported="true"
             android:label="@string/title_activity_welcome"
             android:theme="@style/Theme.Polypoly">
